airport/requirements.txt
asgiref==3.4.1
attrs==21.2.0
coverage==5.5
Django==3.2.7
django-debug-toolbar==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.8.1
iniconfig==1.1.1
Markdown==3.3.4
packaging==21.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.5.0
